Apply assess_* family of functions to a package reference

By default, use all assess_* funtions in the riskmetric namespace and produce a tibble with one column per assessment applied.

Usage

pkg_assess(
  x,
  assessments = all_assessments(),
  ...,
  error_handler = assessment_error_empty
)

Arguments

x

A single pkg_ref object or tibble of package references to assess

assessments

A list of assessment functions to apply to each package reference. By default, a list of all exported assess_* functions from the riskmetric package.

...

additional arguments unused

error_handler

A function, which accepts a single parameter expecting the raised error, which will be called if any errors occur when attempting to apply an assessment function.

Value

Either a list_of_pkg_metric object when a single

pkg_ref object is passed as x, or a

tibble of metrics when a list_of_pkg_ref or

tibble is passed as x. When a tibble

is returned, it has one row per package reference and a new column per assessment function, with cells of that column as package metric objects returned when the assessment was called with the associated pacakge reference.

Assessment function catalog

assess_last_30_bugs_status

vector indicating whether BugReports status is closed

assess_covr_coverage

Package unit test coverage

assess_size_codebase

number of lines of code base

assess_export_help

exported objects have documentation

assess_r_cmd_check

Package check results

assess_dependencies

Package dependency footprint

assess_reverse_dependencies

List of reverse dependencies a package has

assess_license

software is released with an acceptable license

assess_has_maintainer

a vector of associated maintainers

assess_remote_checks

Number of OS flavors that passed/warned/errored on R CMD check

assess_exported_namespace

Objects exported by package

assess_has_website

a vector of associated website urls

assess_downloads_1yr

number of downloads in the past year

assess_has_news

number of discovered NEWS files

assess_has_vignettes

number of discovered vignettes files

assess_has_examples

proportion of discovered function files with examples

assess_has_source_control

a vector of associated source control urls

assess_has_bug_reports_url

presence of a bug reports url in repository

assess_news_current

NEWS file contains entry for current version number